create separate test file for truncated bptt
